refactor: 重构目录结构
