+++ src/main.ts
@@ -1,16 +1,11 @@
-import { Dialog } from 'vant'
-
 import App from './App.vue'
 import router from './router'
 import pinia from '@/plugins/pinia'
 import '@/assets/style/tailwind.css'
 import '@vant/touch-emulator'
 
-// 全局注册
-
 const app = createApp(App)
 
-// app.use(Dialog)
 app.use(router)
 app.use(pinia)
 app.mount('#app')
