+++ src/views/attribute-panel/components/components/IndefiniteNumberInputBox.module.scss
