+++ src/views/attribute-panel/components/components/SwitchWithSlots.tsx
