+++ src/views/attribute-panel/config.ts
