+++ src/views/edit-panel/config.ts
@@ -1,6 +1,6 @@
 import type { Draggable } from '@/components/base-ui/kzy-draggable/types'
 
-export const editDraggableConfig: Draggable = {
+export const config: Draggable = {
   draggableProp: {
     group: { name: 'library' },
     itemKey: 'id',
