+++ src/views/home/components/code-tab.vue
@@ -5,7 +5,11 @@
 <script lang="ts" setup>
 import { onMounted, ref, shallowRef, watch } from 'vue'
 import * as Monaco from 'monaco-editor'
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
 import EditorWorker from 'monaco-editor/esm/vs/editor/editor.worker?worker'
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
 import JsonWorker from 'monaco-editor/esm/vs/language/json/json.worker?worker'
 import { throttle } from 'lodash-es'
 import { useCodeStore } from '@/stores/code'
@@ -51,13 +55,13 @@ onMounted(() => {
   /**
    * An event emitted when the content of the current model has changed.
    */
-  editorInstanceRef.value.onDidChangeModelContent((e) => {
+  editorInstanceRef.value.onDidChangeModelContent(() => {
     formatCode()
   })
   /**
    * 第一次点开时候虽然调用了，但是Monaco并没有相应格式化操作，所以这里延迟一下
    */
-  setTimeout(() => formatCode(), 50)
+  setTimeout(() => formatCode(), 100)
   editorTextModel = editorInstanceRef.value.getModel()
   watch(
     jsonCode,
