+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "@vue/tsconfig/tsconfig.web.json",
-  "include": ["src/**/*", "src/**/*.vue", "*.d.ts"],
+  "include": ["src/**/*", "src/**/*.vue", "*.d.ts", "types/*.d.ts"],
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
